requirements.txt
git+https://github.com/m-bain/whisperx.git@main
ffmpeg-python